+++ test/index.js
@@ -39,9 +39,11 @@ describe('markdown-it-link-attributes', function () {
     expect(result).to.contain('<a href="https://google.com" target="_blank" rel="noopener" foo="bar">link</a>')
   })
 
-  it('takes pattern option and only apply attrs if pattern matched', function () {
+  it('takes matcher function if it returns true', function () {
     this.md.use(linkAttributes, {
-      pattern: /^https?:\/\//,
+      matcher (href) {
+        return /^https?:\/\//.test(href)
+      },
       attrs: {
         target: '_blank',
         rel: 'noopener'
@@ -55,29 +57,18 @@ describe('markdown-it-link-attributes', function () {
     expect(result).to.contain('<a href="#anchor">link</a>')
   })
 
-  it('allows pattern passed to be a string representation of a RegExp', function () {
-    this.md.use(linkAttributes, {
-      pattern: '^https',
-      attrs: {
-        target: '_blank'
-      }
-    })
-
-    var result = this.md.render('[link](https://google.com)')
-    expect(result).to.contain('<a href="https://google.com" target="_blank">link</a>')
-
-    result = this.md.render('[link](#anchor)')
-    expect(result).to.contain('<a href="#anchor">link</a>')
-  })
-
   it('allows multiple rules', function () {
     this.md.use(linkAttributes, [{
-      pattern: /^https:/,
+      matcher (href) {
+        return href.indexOf('https://') === 0
+      },
       attrs: {
         class: 'has-text-uppercase'
       }
     }, {
-      pattern: /^#/,
+      matcher (href) {
+        return href.indexOf('#') === 0
+      },
       attrs: {
         class: 'is-blue'
       }
@@ -95,6 +86,86 @@ describe('markdown-it-link-attributes', function () {
 
     result = this.md.render('[About](/page/about)')
     expect(result).to.contain('<a href="/page/about" class="is-red">About</a>')
+  })
+
+  it('uses the first rule that matches if multiple match', function () {
+    this.md.use(linkAttributes, [{
+      matcher (href) {
+        return href.includes('g')
+      },
+      attrs: {
+        class: 'contains-g'
+      }
+    }, {
+      matcher (href) {
+        return href.indexOf('https://') === 0
+      },
+      attrs: {
+        class: 'starts-with-https'
+      }
+    }, {
+      matcher (href) {
+        return href.indexOf('http') === 0
+      },
+      attrs: {
+        class: 'starts-with-http'
+      }
+    }])
+
+    var result = this.md.render('[Google](https://www.google.com)')
+    expect(result).to.contain('<a href="https://www.google.com" class="contains-g">Google</a>')
+
+    result = this.md.render('[Not Google](https://www.example.com)')
+    expect(result).to.contain('<a href="https://www.example.com" class="starts-with-https">Not Google</a>')
+
+    result = this.md.render('[Not Google and not secure](http://www.example.com)')
+    expect(result).to.contain('<a href="http://www.example.com" class="starts-with-http">Not Google and not secure</a>')
+
+    result = this.md.render('[Not Google and not secure](http://www.example.com/g)')
+    expect(result).to.contain('<a href="http://www.example.com/g" class="contains-g">Not Google and not secure</a>')
+  })
+
+  // NEXT_MAJOR_VERSION we should probably apply all that apply instead of just going with the first to apply
+  // The problem will be when multiple attrs are modifying the same property, in which case we'll probably just want to go with the first
+  it('only uses the first rule if the first rule has no matcher', function () {
+    this.md.use(linkAttributes, [{
+      attrs: {
+        class: 'always-use-this'
+      }
+    }, {
+      matcher (href) {
+        return href.includes('g')
+      },
+      attrs: {
+        class: 'contains-g'
+      }
+    }, {
+      matcher (href) {
+        return href.indexOf('https://') === 0
+      },
+      attrs: {
+        class: 'starts-with-https'
+      }
+    }, {
+      matcher (href) {
+        return href.indexOf('http') === 0
+      },
+      attrs: {
+        class: 'starts-with-http'
+      }
+    }])
+
+    var result = this.md.render('[Google](https://www.google.com)')
+    expect(result).to.contain('<a href="https://www.google.com" class="always-use-this">Google</a>')
+
+    result = this.md.render('[Not Google](https://www.example.com)')
+    expect(result).to.contain('<a href="https://www.example.com" class="always-use-this">Not Google</a>')
+
+    result = this.md.render('[Not Google and not secure](http://www.example.com)')
+    expect(result).to.contain('<a href="http://www.example.com" class="always-use-this">Not Google and not secure</a>')
+
+    result = this.md.render('[Not Google and not secure](http://www.example.com/g)')
+    expect(result).to.contain('<a href="http://www.example.com/g" class="always-use-this">Not Google and not secure</a>')
   })
 
   it('treats className as if it is class', function () {