+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# unreleased
+# 5.0.0
+
+* Add `matcher` function configuration
+
+_Breaking Changes_
 
 * Support latest major version of Markdown-it (currently 12)
 * Drop Node < v14
+* Drop `pattern` configuration